Switching to .NET 5.0
Replacing with work done on a copy to experiment with porting to .NET core.
Much of the Designer still needs work. but the forms open now.
Updating primitive structs to work split up into more focused libraries.
Did a little work on Midi project. Still needs a lot of work.
Figured out a way to split media files in different dlls and register them with the main library.
Cleaned some of the code.

+++ Engine.Base/Framework/ArrayUtilities.cs
@@ -30,8 +30,11 @@ namespace Engine
         /// </remarks>
         public static void Add<T>(ref T[] array, T t)
         {
-            Array.Resize(ref array, array.Length + 1);
-            array[array.Length - 1] = t;
+            if (!(array is null))
+            {
+                Array.Resize(ref array, (array?.Length).Value + 1);
+                array[^1] = t;
+            }
         }
 
         /// <summary>
@@ -42,23 +45,24 @@ namespace Engine
         /// <typeparam name="T"></typeparam>
         public static void RemoveAt<T>(ref T[] array, int index)
         {
-            if (index < array.Length - 1)
+            if (!(array is null))
             {
-                Array.Copy(array, index + 1, array, index, array.Length - index - 1);
-            }
+                if (index < array.Length - 1)
+                {
+                    Array.Copy(array, index + 1, array, index, array.Length - index - 1);
+                }
 
-            Array.Resize(ref array, array.Length - 1);
+                Array.Resize(ref array, array.Length - 1);
+            }
         }
 
         /// <summary>
         /// Remove the at.
         /// </summary>
-        /// <typeparam name="T"></typeparam>
         /// <param name="array">The array.</param>
         /// <param name="index">The index.</param>
-        /// <returns>
-        /// The <see cref="Array" />.
-        /// </returns>
+        /// <returns>The <see cref="Array"/>.</returns>
+        /// <typeparam name="T"></typeparam>
         public static T[] RemoveAt<T>(this T[] array, int index)
         {
             array.RemoveAt(index);
@@ -73,9 +77,14 @@ namespace Engine
         /// <typeparam name="T"></typeparam>
         public static T Pop<T>(this List<T> list)
         {
-            var r = list[list.Count - 1];
-            list.RemoveAt(list.Count - 1);
-            return r;
+            if (!(list is null))
+            {
+                var r = list[^1];
+                list.RemoveAt(list.Count - 1);
+                return r;
+            }
+
+            return default;
         }
 
         /// <summary>
@@ -86,9 +95,14 @@ namespace Engine
         /// <typeparam name="T"></typeparam>
         public static T Shift<T>(this List<T> list)
         {
-            var r = list[0];
-            list.RemoveAt(0);
-            return r;
+            if (!(list is null))
+            {
+                var r = list[0];
+                list.RemoveAt(0);
+                return r;
+            }
+
+            return default;
         }
 
         /// <summary>
@@ -100,7 +114,11 @@ namespace Engine
         /// <typeparam name="T"></typeparam>
         public static List<T> UnShift<T>(this List<T> list, T item)
         {
-            list.Insert(0, item);
+            if (!(list is null))
+            {
+                list.Insert(0, item);
+            }
+
             return list;
         }
 
@@ -108,55 +126,61 @@ namespace Engine
         /// Get the array slice between the two indexes.
         /// ... Inclusive for start index, exclusive for end index.
         /// </summary>
-        /// <typeparam name="T"></typeparam>
         /// <param name="source">The source.</param>
         /// <param name="start">The start.</param>
         /// <param name="end">The end.</param>
-        /// <returns>
-        /// The <see cref="Array" />.
-        /// </returns>
+        /// <returns>The <see cref="Array"/>.</returns>
+        /// <typeparam name="T"></typeparam>
         /// <remarks>
         /// <para>https://www.dotnetperls.com/array-slice</para>
         /// </remarks>
         public static T[] Slice<T>(this T[] source, int start, int end)
         {
-            // Handles negative ends.
-            if (end < 0)
+            if (!(source is null))
             {
-                end = source.Length + end;
-            }
-            var len = end - start;
+                // Handles negative ends.
+                if (end < 0)
+                {
+                    end = source.Length + end;
+                }
+                var len = end - start;
 
-            // Return new array.
-            var res = new T[len];
-            for (var i = 0; i < len; i++)
-            {
-                res[i] = source[i + start];
+                // Return new array.
+                var res = new T[len];
+                for (var i = 0; i < len; i++)
+                {
+                    res[i] = source[i + start];
+                }
+                return res;
             }
-            return res;
+
+            return source;
         }
 
         /// <summary>
         /// Get the array slice between the two indexes.
         /// ... Inclusive for start index, exclusive for end index.
         /// </summary>
-        /// <typeparam name="T"></typeparam>
         /// <param name="source">The source.</param>
-        /// <returns>
-        /// The <see cref="Array" />.
-        /// </returns>
+        /// <returns>The <see cref="Array"/>.</returns>
+        /// <typeparam name="T"></typeparam>
         /// <remarks>
         /// <para>https://www.dotnetperls.com/array-slice</para>
         /// </remarks>
         public static T[] Slice<T>(this T[] source)
         {
-            // Return new array.
-            var res = new T[source.Length];
-            for (var i = 0; i < source.Length; i++)
+            if (!(source is null))
             {
-                res[i] = source[i];
+                // Return new array.
+                var res = new T[source.Length];
+                for (var i = 0; i < source.Length; i++)
+                {
+                    res[i] = source[i];
+                }
+                return res;
             }
-            return res;
+
+            return source;
         }
 
         /// <summary>
@@ -172,9 +196,12 @@ namespace Engine
         /// </remarks>
         public static IEnumerable<TResult> Map<T, TResult>(this IEnumerable<T> list, Func<T, TResult> func)
         {
-            foreach (var i in list)
+            if (!(list is null))
             {
-                yield return func(i);
+                foreach (var i in list)
+                {
+                    yield return func(i);
+                }
             }
         }
 
@@ -192,9 +219,12 @@ namespace Engine
         /// </remarks>
         public static T Reduce<T, U>(this IEnumerable<U> list, Func<U, T, T> func, T acc)
         {
-            foreach (var i in list)
+            if (!(list is null))
             {
-                acc = func(i, acc);
+                foreach (var i in list)
+                {
+                    acc = func(i, acc);
+                }
             }
 
             return acc;
@@ -213,9 +243,14 @@ namespace Engine
         /// </remarks>
         public static List<T> Splice<T>(this List<T> Source, int Start, int Size)
         {
-            var retVal = Source.Skip(Start).Take(Size).ToList();
-            Source.RemoveRange(Start, Size);
-            return retVal;
+            if (!(Source is null))
+            {
+                var retVal = Source.Skip(Start).Take(Size).ToList();
+                Source.RemoveRange(Start, Size);
+                return retVal;
+            }
+
+            return Source;
         }
 
         /// <summary>
@@ -231,8 +266,10 @@ namespace Engine
         {
             var randomList = new List<T>();
 
+#pragma warning disable SecurityIntelliSenseCS // MS Security rules violation
             var r = new Random();
-            while (inputList.Count > 0)
+#pragma warning restore SecurityIntelliSenseCS // MS Security rules violation
+            while (inputList?.Count > 0)
             {
                 //Choose a random object in the list
                 var randomIndex = r.Next(0, inputList.Count);
